requirements.txt
requests~=2.32.0
numpy~=2.2.0
pillow~=10.4.0
